+++ syncany-lib/src/main/java/org/syncany/plugins/local/LocalTransferManager.java
@@ -19,7 +19,10 @@ package org.syncany.plugins.local;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.HashMap;
+import java.nio.file.DirectoryStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -40,6 +43,8 @@ import org.syncany.plugins.transfer.files.SyncanyRemoteFile;
 import org.syncany.plugins.transfer.files.TempRemoteFile;
 import org.syncany.plugins.transfer.files.TransactionRemoteFile;
 
+import com.google.common.collect.Maps;
+
 /**
  * Implements a {@link TransferManager} based on a local storage backend for the
  * {@link LocalTransferPlugin}.
@@ -50,8 +55,8 @@ import org.syncany.plugins.transfer.files.TransactionRemoteFile;
  * in special sub-folders:
  *
  * <ul>
- *   <li>The <tt>databases</tt> folder keeps all the {@link DatabaseRemoteFile}s</li>
- *   <li>The <tt>multichunks</tt> folder keeps the actual data within the {@link MultichunkRemoteFile}s</li>
+ * <li>The <tt>databases</tt> folder keeps all the {@link DatabaseRemoteFile}s</li>
+ * <li>The <tt>multichunks</tt> folder keeps the actual data within the {@link MultichunkRemoteFile}s</li>
  * </ul>
  *
  * <p>This plugin can be used for testing or to point to a repository
@@ -63,22 +68,22 @@ import org.syncany.plugins.transfer.files.TransactionRemoteFile;
 public class LocalTransferManager extends AbstractTransferManager {
 	private static final Logger logger = Logger.getLogger(LocalTransferManager.class.getSimpleName());
 
-	private File repoPath;
-	private File multichunksPath;
-	private File databasesPath;
-	private File actionsPath;
-	private File transactionsPath;
-	private File temporaryPath;
+	private Path repoPath;
+	private Path multichunksPath;
+	private Path databasesPath;
+	private Path actionsPath;
+	private Path transactionsPath;
+	private Path temporaryPath;
 
 	public LocalTransferManager(LocalTransferSettings connection, Config config) {
 		super(connection, config);
 
-		this.repoPath = connection.getPath().getAbsoluteFile(); // absolute file to get abs. path!
-		this.multichunksPath = new File(connection.getPath().getAbsolutePath(), "multichunks");
-		this.databasesPath = new File(connection.getPath().getAbsolutePath(), "databases");
-		this.actionsPath = new File(connection.getPath().getAbsolutePath(), "actions");
-		this.transactionsPath = new File(connection.getPath().getAbsolutePath(), "transactions");
-		this.temporaryPath = new File(connection.getPath().getAbsolutePath(), "temporary");
+		this.repoPath = Paths.get(connection.getPath().toURI()); // absolute file to get abs. path!
+		this.multichunksPath = repoPath.resolve("multichunks");
+		this.databasesPath = repoPath.resolve("databases");
+		this.actionsPath = repoPath.resolve("actions");
+		this.transactionsPath = repoPath.resolve("transactions");
+		this.temporaryPath = repoPath.resolve("temporary");
 	}
 
 	@Override
@@ -97,30 +102,35 @@ public class LocalTransferManager extends AbstractTransferManager {
 	public void init(boolean createIfRequired) throws StorageException {
 		connect();
 
-		if (!testTargetExists() && createIfRequired) {
-			if (!repoPath.mkdir()) {
-				throw new StorageException("Cannot create repository directory: " + repoPath);
+		try {
+			if (!testTargetExists() && createIfRequired) {
+				if (!Files.exists(Files.createDirectory(repoPath))) {
+					throw new StorageException("Cannot create repository directory: " + repoPath);
+				}
+			}
+
+			if (!Files.exists(Files.createDirectory(multichunksPath))) {
+				throw new StorageException("Cannot create multichunk directory: " + multichunksPath);
+			}
+
+			if (!Files.exists(Files.createDirectory(databasesPath))) {
+				throw new StorageException("Cannot create databases directory: " + databasesPath);
+			}
+
+			if (!Files.exists(Files.createDirectory(actionsPath))) {
+				throw new StorageException("Cannot create actions directory: " + actionsPath);
+			}
+
+			if (!Files.exists(Files.createDirectory(transactionsPath))) {
+				throw new StorageException("Cannot create transactions directory: " + transactionsPath);
+			}
+
+			if (!Files.exists(Files.createDirectory(temporaryPath))) {
+				throw new StorageException("Cannot create temporary directory: " + temporaryPath);
 			}
 		}
-
-		if (!multichunksPath.mkdir()) {
-			throw new StorageException("Cannot create multichunk directory: " + multichunksPath);
-		}
-
-		if (!databasesPath.mkdir()) {
-			throw new StorageException("Cannot create databases directory: " + databasesPath);
-		}
-
-		if (!actionsPath.mkdir()) {
-			throw new StorageException("Cannot create actions directory: " + actionsPath);
-		}
-
-		if (!transactionsPath.mkdir()) {
-			throw new StorageException("Cannot create transactions directory: " + transactionsPath);
-		}
-
-		if (!temporaryPath.mkdir()) {
-			throw new StorageException("Cannot create temporary directory: " + temporaryPath);
+		catch (IOException e) {
+			throw new StorageException("Unable to create directories", e);
 		}
 	}
 
@@ -200,66 +210,61 @@ public class LocalTransferManager extends AbstractTransferManager {
 
 		File repoFile = getRemoteFile(remoteFile);
 
-		if (!repoFile.exists()) {
-			return true;
-		}
+		return !repoFile.exists() || repoFile.delete();
 
-		return repoFile.delete();
 	}
 
 	@Override
 	public <T extends RemoteFile> Map<String, T> list(Class<T> remoteFileClass) throws StorageException {
 		connect();
 
-		// List folder
-		File remoteFilePath = getRemoteFilePath(remoteFileClass);
-		File[] files = remoteFilePath.listFiles();
+		Path folder = Paths.get(getRemoteFilePath(remoteFileClass));
+		Map<String, T> files = Maps.newHashMap();
 
-		if (files == null) {
-			logger.log(Level.FINE, remoteFilePath.getAbsolutePath());
-			throw new StorageException("Unable to read local respository " + repoPath);
-		}
-
-		// Create RemoteFile objects
-		Map<String, T> remoteFiles = new HashMap<String, T>();
-
-		for (File file : files) {
-			try {
-				T remoteFile = RemoteFile.createRemoteFile(file.getName(), remoteFileClass);
-				remoteFiles.put(file.getName(), remoteFile);
-			}
-			catch (StorageException e) {
-				logger.log(Level.INFO, "Cannot create instance of " + remoteFileClass.getSimpleName() + " for file " + file
-						+ "; maybe invalid file name pattern. Ignoring file.");
+		try (DirectoryStream<Path> directoryStream = Files.newDirectoryStream(folder)) {
+			for (Path path : directoryStream) {
+				try {
+					T remoteFile = RemoteFile.createRemoteFile(path.getFileName().toString(), remoteFileClass);
+					files.put(path.getFileName().toString(), remoteFile);
+				}
+				catch (StorageException e) {
+					logger.log(Level.INFO, "Cannot create instance of " + remoteFileClass.getSimpleName() + " for file " + path
+									+ "; maybe invalid file name pattern. Ignoring file.", e);
+				}
 			}
 		}
+		catch (IOException e) {
+			logger.log(Level.SEVERE, "Unable to list directory", e);
+		}
 
-		return remoteFiles;
+		return files;
+	}
+
+	@Override
+	public String getRemoteFilePath(Class<? extends RemoteFile> remoteFile) {
+		if (remoteFile.equals(MultichunkRemoteFile.class)) {
+			return multichunksPath.toString();
+		}
+		else if (remoteFile.equals(DatabaseRemoteFile.class) || remoteFile.equals(CleanupRemoteFile.class)) {
+			return databasesPath.toString();
+		}
+		else if (remoteFile.equals(ActionRemoteFile.class)) {
+			return actionsPath.toString();
+		}
+		else if (remoteFile.equals(TransactionRemoteFile.class)) {
+			return transactionsPath.toString();
+		}
+		else if (remoteFile.equals(TempRemoteFile.class)) {
+			return temporaryPath.toString();
+		}
+		else {
+			return repoPath.toString();
+		}
 	}
 
 	private File getRemoteFile(RemoteFile remoteFile) {
-		return new File(getRemoteFilePath(remoteFile.getClass()), remoteFile.getName());
-	}
-
-	private File getRemoteFilePath(Class<? extends RemoteFile> remoteFile) {
-		if (remoteFile.equals(MultichunkRemoteFile.class)) {
-			return multichunksPath;
-		}
-		else if (remoteFile.equals(DatabaseRemoteFile.class) || remoteFile.equals(CleanupRemoteFile.class)) {
-			return databasesPath;
-		}
-		else if (remoteFile.equals(ActionRemoteFile.class)) {
-			return actionsPath;
-		}
-		else if (remoteFile.equals(TransactionRemoteFile.class)) {
-			return transactionsPath;
-		}
-		else if (remoteFile.equals(TempRemoteFile.class)) {
-			return temporaryPath;
-		}
-		else {
-			return repoPath;
-		}
+		String rootPath = getRemoteFilePath(remoteFile.getClass());
+		return Paths.get(rootPath, remoteFile.getName()).toFile();
 	}
 
 	public String getAbsoluteParentDirectory(File file) {
@@ -269,9 +274,9 @@ public class LocalTransferManager extends AbstractTransferManager {
 	@Override
 	public boolean testTargetCanWrite() {
 		try {
-			if (repoPath.isDirectory()) {
-				File tempFile = File.createTempFile("syncany-write-test", "tmp", repoPath);
-				tempFile.delete();
+			if (Files.isDirectory(repoPath)) {
+				Path tempFile = Files.createTempFile(repoPath, "syncany-write-test", "tmp");
+				Files.delete(tempFile);
 
 				logger.log(Level.INFO, "testTargetCanWrite: Can write, test file created/deleted successfully.");
 				return true;
@@ -289,7 +294,7 @@ public class LocalTransferManager extends AbstractTransferManager {
 
 	@Override
 	public boolean testTargetExists() {
-		if (repoPath.exists()) {
+		if (Files.exists(repoPath)) {
 			logger.log(Level.INFO, "testTargetExists: Target exists.");
 			return true;
 		}
@@ -321,7 +326,7 @@ public class LocalTransferManager extends AbstractTransferManager {
 
 	@Override
 	public boolean testTargetCanCreate() {
-		if (repoPath.getParentFile().canWrite()) {
+		if (Files.isWritable(repoPath.getParent())) {
 			logger.log(Level.INFO, "testTargetCanCreate: Can create target.");
 			return true;
 		}